Refactore location of response objects
